Add BridgeApp Swift package
