+++ Cartfile
@@ -1,2 +1,2 @@
-github "Sage-Bionetworks/SageResearch" ~> 3.9
-github "Sage-Bionetworks/Bridge-iOS-SDK" ~> 4.2.72
+github "Sage-Bionetworks/SageResearch" ~> 3.17
+github "Sage-Bionetworks/Bridge-iOS-SDK" ~> 4.4
