+++ Cartfile.resolved
@@ -1,2 +1,2 @@
-github "Sage-Bionetworks/Bridge-iOS-SDK" "v4.3.76"
-github "Sage-Bionetworks/SageResearch" "3.14.3"
+github "Sage-Bionetworks/Bridge-iOS-SDK" "4.4.77"
+github "Sage-Bionetworks/SageResearch" "3.17.1"
