+++ Package.swift
@@ -18,135 +18,71 @@ let package = Package(
         .library(
             name: "BridgeApp",
             targets: ["BridgeApp"]),
-//        .library(
-//            name: "BridgeAppUI",
-//            targets: ["BridgeAppUI"]),
-//        .library(
-//            name: "DataTracking",
-//            targets: ["DataTracking"]),
+        .library(
+            name: "BridgeAppUI",
+            targets: ["BridgeAppUI"]),
+        .library(
+            name: "DataTracking",
+            targets: ["DataTracking"]),
 
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         .package(name: "SageResearch",
                  url: "https://github.com/Sage-Bionetworks/SageResearch.git",
-                 from: "3.14"),
+                 from: "3.14.0"),
         .package(name: "BridgeSDK",
-                 url: "https://github.com/syoung-smallwisdom/BridgeSDK.git",
-                 .branch("swiftPM")),
+                 url: "https://github.com/Sage-Bionetworks/Bridge-iOS-SDK.git",
+                 from: "4.4.77"),
         .package(name: "JsonModel",
                  url: "https://github.com/Sage-Bionetworks/JsonModel-Swift.git",
                  from: "1.0.2"),
     ],
     targets: [
 
-        // Research is the main target included in this repo. The "Formatters" and
-        // "ExceptionHandler" targets are developed in Obj-c so they require a
-        // separate target.
         .target(
             name: "BridgeApp",
-            dependencies: ["SageResearch",
-                           "BridgeSDK",
-                           "JsonModel",
+            dependencies: [
+                .product(name: "Research", package: "SageResearch"),
+                .product(name: "ResearchUI", package: "SageResearch"),
+                "BridgeSDK",
+                "JsonModel",
             ],
-            path: "BridgeApp/BridgeApp/",
-            exclude: ["Info-iOS.plist"],
+            path: "BridgeApp/BridgeApp/iOS",
             resources: [
                 .process("Localization"),
             ]
             ),
         
-//        // ResearchUI currently only supports iOS devices. This includes views and view
-//        // controllers and references UIKit.
-//        .target(
-//            name: "ResearchUI",
-//            dependencies: [
-//                "Research",
-//            ],
-//            path: "Research/ResearchUI/",
-//            exclude: ["Info-iOS.plist"],
-//            resources: [
-//                .process("PlatformContext/Resources"),
-//                .process("iOS/Resources"),
-//            ]),
-//
-//        // ResearchAudioRecorder is used to allow recording dbFS level.
-//        .target(
-//            name: "ResearchAudioRecorder",
-//            dependencies: [
-//                "Research",
-//            ],
-//            path: "Research/ResearchAudioRecorder/",
-//            exclude: ["Info.plist"]),
-//
-//        // ResearchMotion is used to allow recording motion sensors.
-//        .target(
-//            name: "ResearchMotion",
-//            dependencies: [
-//                "Research",
-//            ],
-//            path: "Research/ResearchMotion/",
-//            exclude: ["Info.plist"],
-//            resources: [
-//                .process("Resources"),
-//            ]),
-//
-//        // ResearchLocation is used to allow location authorization and record distance
-//        // travelled.
-//        .target(
-//            name: "ResearchLocation",
-//            dependencies: [
-//                "Research",
-//                "ResearchMotion",
-//            ],
-//            path: "Research/ResearchLocation/",
-//            exclude: ["Info.plist"]),
-//
-//        // The following targets are set up for unit testing.
-//        .target(
-//            name: "Research_UnitTest",
-//            dependencies: ["Research",
-//                           "ResearchUI",
-//            ],
-//            path: "Research/Research_UnitTest/",
-//            exclude: ["Info.plist"]),
-//        .target(name: "NSLocaleSwizzle",
-//                dependencies: [],
-//                path: "Research/NSLocaleSwizzle/",
-//                exclude: ["Info.plist"]),
-//        .testTarget(
-//            name: "ResearchTests",
-//            dependencies: [
-//                "Research",
-//                "Research_UnitTest",
-//                "NSLocaleSwizzle",
-//            ],
-//            path: "Research/ResearchTests/",
-//            exclude: ["Info-iOS.plist"],
-//            resources: [
-//                .process("Resources"),
-//            ]),
-//        .testTarget(
-//            name: "ResearchUITests",
-//            dependencies: ["ResearchUI"],
-//            path: "Research/ResearchUITests/",
-//            exclude: ["Info.plist"]),
-//        .testTarget(
-//            name: "ResearchMotionTests",
-//            dependencies: [
-//                "ResearchMotion",
-//                "Research_UnitTest",
-//            ],
-//            path: "Research/ResearchMotionTests/",
-//            exclude: ["Info.plist"]),
-//        .testTarget(
-//            name: "ResearchLocationTests",
-//            dependencies: [
-//                "ResearchLocation",
-//                "Research_UnitTest",
-//            ],
-//            path: "Research/ResearchLocationTests/",
-//            exclude: ["Info.plist"]),
+        .target(
+            name: "BridgeAppUI",
+            dependencies: [
+                .product(name: "Research", package: "SageResearch"),
+                .product(name: "ResearchUI", package: "SageResearch"),
+                "BridgeApp",
+                "BridgeSDK",
+            ],
+            path: "BridgeApp/BridgeAppUI/iOS",
+            resources: [
+                .process("Resources"),
+            ]
+            ),
+        
+        .target(
+            name: "DataTracking",
+            dependencies: [
+                .product(name: "Research", package: "SageResearch"),
+                .product(name: "ResearchUI", package: "SageResearch"),
+                "BridgeApp",
+                "BridgeAppUI",
+                "BridgeSDK",
+                "JsonModel",
+            ],
+            path: "BridgeApp/DataTracking/iOS",
+            resources: [
+                .process("Resources"),
+            ]
+            ),
         
     ]
 )
